pyproject.toml
[project]
name = "fprac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "psycopg2-binary",
    "sqlalchemy==2.0.34", # 👈 look here
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
]
